RNA-Seq
I did a RNA-Seq and a diagrama to demonstrate RNA-Seq processes in a more understable way. I did also two plots (dispersion plot and MA plots), a line to see differentially expressed genes number, and a line to see differentially expressed genes.  

In the dispersion plot red line is the trend line, blue dots represent the genes expression final variability and black dots represent dispersion puntual estimate for each gene.
    
There are three MA plots to demostrate the difference in differentially expressed genes when we change adjusted p-values. When alpha value is lower, the search for differentially expressed genes is stricter. When alpha value is high, result is the opposite.